# contracts/openapi/brivva-workers.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Brivva Workers API",
    "version": "0.2.0",
    "description": "CRUD, auth, and internal session bootstrap API for Brivva."
  },
  "servers": [
    {
      "url": "https://brivva-api.milliytechnology.workers.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "InternalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Secret"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "path",
                "message"
              ],
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Voice": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "elevenlabs_voice_id",
          "name",
          "source_lang",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "elevenlabs_voice_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "source_lang": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "integer"
          }
        }
      },
      "Stream": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "lang",
          "platform",
          "platform_broadcast_id",
          "platform_stream_id",
          "stream_key",
          "rtmp_url",
          "status",
          "delay_ms",
          "host_gain",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "platform_broadcast_id": {
            "type": "string",
            "nullable": true
          },
          "platform_stream_id": {
            "type": "string",
            "nullable": true
          },
          "stream_key": {
            "type": "string",
            "nullable": true
          },
          "rtmp_url": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "delay_ms": {
            "type": "integer"
          },
          "host_gain": {
            "type": "number"
          },
          "created_at": {
            "type": "integer"
          },
          "watch_url": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Session": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "voice_id",
          "title",
          "source_lang",
          "target_langs",
          "status",
          "live_session_id",
          "voice_preset",
          "created_at",
          "translation_terms"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "voice_id": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "source_lang": {
            "type": "string"
          },
          "target_langs": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "live_session_id": {
            "type": "string",
            "nullable": true
          },
          "voice_preset": {
            "type": "string",
            "enum": [
              "cloned",
              "female",
              "male"
            ]
          },
          "created_at": {
            "type": "integer"
          },
          "translation_terms": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ListVoicesResponse": {
        "type": "object",
        "required": [
          "voices"
        ],
        "properties": {
          "voices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Voice"
            }
          }
        }
      },
      "CreateVoiceRequest": {
        "type": "object",
        "required": [
          "user_id",
          "name",
          "audio_base64"
        ],
        "properties": {
          "user_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "audio_base64": {
            "type": "string"
          },
          "source_lang": {
            "type": "string"
          }
        }
      },
...
      "InternalSessionBundle": {
        "type": "object",
        "required": [
          "session",
          "streams",
          "voice"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/Session"
          },
          "streams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stream"
            }
          },
          "voice": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Voice"
              }
            ],
            "nullable": true
          }
        }
      },
      "InternalSessionStatusUpdate": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "live_session_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "InternalSessionStatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ]
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/voices": {
      "get": {
        "summary": "List saved voices",
        "responses": {
          "200": {
            "description": "Voice list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListVoicesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create voice clone",
        "responses": {
          "200": {
            "description": "Created voice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoiceRequest"
              }
            }
          }
        }
      }
    },
...
    "/internal/sessions/{id}": {
      "get": {
        "summary": "Fetch session bootstrap bundle for live session startup",
        "responses": {
          "200": {
            "description": "Session bundle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalSessionBundle"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "InternalSecret": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "Update session status from media backend",
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalSessionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "InternalSecret": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalSessionStatusUpdate"
              }
            }
          }
        }
      }
    },
...
  }
}